refactor: 클럽 지원하기 구버전 api 제거
