+++ backend/src/main/java/moadong/club/payload/dto/ClubActiveFormResult.java
@@ -5,7 +5,6 @@ import lombok.Builder;
 @Builder
 public record ClubActiveFormResult(
         String id,
-        String title,
-        String description
+        String title
 ) {
 }
