+++ backend/src/main/java/moadong/club/payload/dto/ClubActiveFormSlim.java
@@ -3,5 +3,4 @@ package moadong.club.payload.dto;
 public interface ClubActiveFormSlim {
     String getId();
     String getTitle();
-    String getDescription();
 }
